🛠️refactor: useMapAutoReturn hooks Convention
